requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-HTTPAuth
Flask-WTF==0.12
gunicorn==19.4.5
pyowm==2.3.0
python-dateutil==2.4.2
pytz
